refactor(types/uint): require exact-same-concrete-type on all binary ops
Every BaseUint binary operator now uses type(other) is type(self) instead
of isinstance(other, BaseUint). Previously, Uint8(5) + Uint16(10) silently
produced Uint8(15) because the result type defaulted to the LHS. Now any
cross-width or cross-newtype mixing raises TypeError, matching how a
strongly typed language rejects u8 + u16 or Slot + Uint64.

Also in this commit:
- __rpow__ gains the missing mod parameter for 3-arg pow.
- __pow__ uses overloads to express the parent's two-overload return shape.
- Dead _validate_int_operand helper removed (no callers remain).
- Class docstring documents the universal strict-type rule.

Consumers throughout the chain/clock, fork-choice, validator service, and
xmss code that were silently relying on cross-type interop are updated to
convert explicitly via int() or the appropriate newtype constructor.

+++ src/lean_spec/subspecs/xmss/interface.py
@@ -117,7 +117,7 @@ class GeneralizedXmssScheme(StrictBaseModel):
         config = self.config
 
         # Ensure the requested activation range is within the scheme's total supported lifetime.
-        if activation_slot + num_active_slots > config.LIFETIME:
+        if int(activation_slot) + int(num_active_slots) > int(config.LIFETIME):
             raise ValueError("Activation range exceeds the key's lifetime.")
 
         # Generate the random public parameter `P` and the master PRF key.
@@ -330,7 +330,7 @@ class GeneralizedXmssScheme(StrictBaseModel):
         bottom_tree = sk.left_bottom_tree if slot_int < boundary else sk.right_bottom_tree
 
         # Generate the combined authentication path
-        path = combined_path(sk.top_tree, bottom_tree, slot)
+        path = combined_path(sk.top_tree, bottom_tree, Uint64(int(slot)))
 
         # Assemble and return the final signature, which contains:
         # - The OTS,
@@ -385,7 +385,7 @@ class GeneralizedXmssScheme(StrictBaseModel):
         #
         # Return False instead of raising to avoid panic on invalid signatures.
         # The slot is attacker-controlled input.
-        if slot >= self.config.LIFETIME:
+        if int(slot) >= int(self.config.LIFETIME):
             return False
 
         # Re-encode the message using the randomness `rho` from the signature.
